Go to a single hook, with an object based API.
