+++ packages/react-hooks/CHANGELOG.md
@@ -7,7 +7,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## [Unreleased] -->
 
-- Added `useToggle` and `useForcibleToggle` hooks ([#1220](https://github.com/Shopify/quilt/pull/1220))
+- Added `useToggle` hook ([#1220](https://github.com/Shopify/quilt/pull/1220))
 
 ## [1.3.0] - 2019-10-29
