+++ packages/react-hooks/src/hooks/index.ts
@@ -1,4 +1,3 @@
-export {useForcibleToggle} from './forcible-toggle';
 export {useLazyRef} from './lazy-ref';
 export {useMountedRef} from './mounted-ref';
 export {useOnValueChange} from './on-value-change';
