+++ packages/react-hooks/src/hooks/tests/toggle.test.tsx
@@ -5,14 +5,16 @@ import {useToggle} from '../toggle';
 
 describe('useToggle', () => {
   function MockComponent({initialValueIsTrue = false}) {
-    const [value, toggleValue] = useToggle(Boolean(initialValueIsTrue));
+    const {value, toggle, setTrue, setFalse} = useToggle(initialValueIsTrue);
 
     const activeText = value ? 'true' : 'false';
 
     return (
       <>
         <p>Value: {activeText}</p>
-        <button type="button" onClick={toggleValue} />
+        <button type="button" id="toggle" onClick={toggle} />
+        <button type="button" id="forceTrue" onClick={setTrue} />
+        <button type="button" id="forceFalse" onClick={setFalse} />
       </>
     );
   }
@@ -25,14 +27,36 @@ describe('useToggle', () => {
     expect(wrapperInitallyTrue).toContainReactText('Value: true');
   });
 
-  it('toggles the value when the callback is triggered', () => {
+  it('toggles the value when the toggle callback is triggered', () => {
     const wrapper = mount(<MockComponent />);
     expect(wrapper).toContainReactText('Value: false');
 
-    wrapper.find('button').trigger('onClick');
+    wrapper.find('button', {id: 'toggle'}).trigger('onClick');
     expect(wrapper).toContainReactText('Value: true');
 
-    wrapper.find('button').trigger('onClick');
+    wrapper.find('button', {id: 'toggle'}).trigger('onClick');
+    expect(wrapper).toContainReactText('Value: false');
+  });
+
+  it('forces the value to true when the forceTrue callback is triggered', () => {
+    const wrapper = mount(<MockComponent />);
+    expect(wrapper).toContainReactText('Value: false');
+
+    wrapper.find('button', {id: 'forceTrue'}).trigger('onClick');
+    expect(wrapper).toContainReactText('Value: true');
+
+    wrapper.find('button', {id: 'forceTrue'}).trigger('onClick');
+    expect(wrapper).toContainReactText('Value: true');
+  });
+
+  it('forces the value to false when the forceFalse callback is triggered', () => {
+    const wrapper = mount(<MockComponent initialValueIsTrue />);
+    expect(wrapper).toContainReactText('Value: true');
+
+    wrapper.find('button', {id: 'forceFalse'}).trigger('onClick');
+    expect(wrapper).toContainReactText('Value: false');
+
+    wrapper.find('button', {id: 'forceFalse'}).trigger('onClick');
     expect(wrapper).toContainReactText('Value: false');
   });
 });
