+++ packages/react-hooks/src/hooks/toggle.ts
@@ -1,13 +1,16 @@
 import {useState, useCallback} from 'react';
 
 /**
- * Returns a stateful value, and a memoized function to toggle it
+ * Returns a stateful value, and a set of memoized functions to toggle it,
+ * set it to true and set it to false
  */
 export function useToggle(initialState: boolean) {
-  const [state, setState] = useState(initialState);
-  const toggle = useCallback(() => setState(state => !state), []);
+  const [value, setState] = useState(initialState);
 
-  // cast needed to say this returns a two item array with the items in
-  // their specific positions instead of `(typeof state | typeof toggle)[]`
-  return [state, toggle] as [typeof state, typeof toggle];
+  return {
+    value,
+    toggle: useCallback(() => setState(state => !state), []),
+    setTrue: useCallback(() => setState(true), []),
+    setFalse: useCallback(() => setState(false), []),
+  };
 }
